+++ Source/Text Fields/TextFieldCell.swift
@@ -1,5 +1,7 @@
 import UIKit
 
+@objc(SDCTextFieldCell)
+@available(iOSApplicationExtension, unavailable)
 final class TextFieldCell: UITableViewCell {
 
     @IBOutlet private var borderView: UIView!
@@ -18,11 +20,10 @@ final class TextFieldCell: UITableViewCell {
         didSet {
             self.textField?.font = self.visualStyle?.textFieldFont
             self.borderView.backgroundColor = self.visualStyle?.textFieldBorderColor
-
+            self.textFieldContainer.backgroundColor = self.visualStyle?.textFieldBackgroundColor
             guard let padding = self.visualStyle?.textFieldMargins else {
                 return
             }
-
             self.paddingConstraints?.leading.constant = padding.left
             self.paddingConstraints?.trailing.constant = -padding.right
             self.paddingConstraints?.top.constant = padding.top
@@ -38,7 +39,7 @@ final class TextFieldCell: UITableViewCell {
         container?.addSubview(textField)
         textField.translatesAutoresizingMaskIntoConstraints = false
 
-        let insets = self.visualStyle?.textFieldMargins ?? UIEdgeInsets.zero
+        let insets = self.visualStyle?.textFieldMargins ?? UIEdgeInsets()
 
         let leading = textField.leadingAnchor.constraint(equalTo: self.leadingAnchor, constant: insets.left)
         let trailing = textField.trailingAnchor.constraint(equalTo: self.trailingAnchor,