+++ Source/Views/AlertView.swift
@@ -1,11 +1,25 @@
-final class AlertView: UIView, AlertControllerViewRepresentable {
+import UIKit
 
-    var titleLabel: AlertLabel! = AlertLabel()
-    var messageLabel: AlertLabel! = AlertLabel()
-    var actionsCollectionView: ActionsCollectionView! = ActionsCollectionView()
-    var contentView: UIView! = UIView()
+@available(iOSApplicationExtension, unavailable)
+final class AlertView: UIView, AlertControllerViewRepresentable {
+    private let scrollView = UIScrollView()
+    private let titleLabel = AlertLabel()
+    private let  messageLabel = AlertLabel()
+    private let actionsCollectionView = ActionsCollectionView()
+
+    let contentView = UIView()
     var actions: [AlertAction] = []
     var actionLayout = ActionLayout.automatic
+
+    var title: NSAttributedString? {
+        get { return self.titleLabel.attributedText }
+        set { self.titleLabel.attributedText = newValue }
+    }
+
+    var message: NSAttributedString? {
+        get { return self.messageLabel.attributedText }
+        set { self.messageLabel.attributedText = newValue }
+    }
 
     var textFieldsViewController: TextFieldsViewController? {
         didSet { self.textFieldsViewController?.visualStyle = self.visualStyle }
@@ -20,11 +34,12 @@ final class AlertView: UIView, AlertControllerViewRepresentable {
         set { self.actionsCollectionView.actionTapped = newValue }
     }
 
-    var topView: UIView {
-        return self.scrollView
-    }
+    var topView: UIView { self.scrollView }
 
-    private let scrollView = UIScrollView()
+    override var intrinsicContentSize: CGSize {
+        let totalHeight = self.contentHeight + self.actionsCollectionView.displayHeight
+        return CGSize(width: UIView.noIntrinsicMetric, height: totalHeight)
+    }
 
     private var elements: [UIView] {
         let possibleElements: [UIView?] = [
@@ -34,11 +49,7 @@ final class AlertView: UIView, AlertControllerViewRepresentable {
             self.contentView.subviews.count > 0 ? self.contentView : nil,
         ]
 
-        #if swift(>=4.1)
-            return possibleElements.compactMap { $0 }
-        #else
-            return possibleElements.flatMap { $0 }
-        #endif
+        return possibleElements.compactMap { $0 }
     }
 
     private var contentHeight: CGFloat {
@@ -71,9 +82,12 @@ final class AlertView: UIView, AlertControllerViewRepresentable {
         self.updateUI()
     }
 
-    func addDragTapBehavior() {
-        let panGesture = UIPanGestureRecognizer(target: self,
-                                                action: #selector(self.highlightAction(for:)))
+    func add(_ behaviors: AlertBehaviors) {
+        if !behaviors.contains(.dragTap) {
+            return
+        }
+
+        let panGesture = UIPanGestureRecognizer(target: self, action: #selector(self.highlightAction(for:)))
         self.addGestureRecognizer(panGesture)
     }
 
@@ -82,16 +96,23 @@ final class AlertView: UIView, AlertControllerViewRepresentable {
     private func createBackground() {
         if let color = self.visualStyle.backgroundColor {
             self.backgroundColor = color
-        } else {
-            let backgroundView = UIVisualEffectView(effect: UIBlurEffect(style: .extraLight))
-            backgroundView.translatesAutoresizingMaskIntoConstraints = false
-
-            self.insertSubview(backgroundView, belowSubview: self.scrollView)
-            backgroundView.leadingAnchor.constraint(equalTo: self.leadingAnchor).isActive = true
-            backgroundView.trailingAnchor.constraint(equalTo: self.trailingAnchor).isActive = true
-            backgroundView.topAnchor.constraint(equalTo: self.topAnchor).isActive = true
-            backgroundView.bottomAnchor.constraint(equalTo: self.bottomAnchor).isActive = true
+            return
         }
+
+        var style: UIBlurEffect.Style
+        if #available(iOS 13.0, *) {
+            style = .systemMaterial
+        } else {
+            style = .extraLight
+        }
+        let backgroundView = UIVisualEffectView(effect: UIBlurEffect(style: style))
+        backgroundView.translatesAutoresizingMaskIntoConstraints = false
+
+        self.insertSubview(backgroundView, belowSubview: self.scrollView)
+        backgroundView.leadingAnchor.constraint(equalTo: self.leadingAnchor).isActive = true
+        backgroundView.trailingAnchor.constraint(equalTo: self.trailingAnchor).isActive = true
+        backgroundView.topAnchor.constraint(equalTo: self.topAnchor).isActive = true
+        backgroundView.bottomAnchor.constraint(equalTo: self.bottomAnchor).isActive = true
     }
 
     private func createUI() {
@@ -120,11 +141,6 @@ final class AlertView: UIView, AlertControllerViewRepresentable {
         self.textFieldsViewController?.visualStyle = self.visualStyle
     }
 
-    override var intrinsicContentSize: CGSize {
-        let totalHeight = self.contentHeight + self.actionsCollectionView.displayHeight
-        return CGSize(width: UIViewNoIntrinsicMetric, height: totalHeight)
-    }
-
     @objc
     private func highlightAction(for sender: UIPanGestureRecognizer) {
         self.actionsCollectionView.highlightAction(for: sender)
@@ -151,9 +167,6 @@ final class AlertView: UIView, AlertControllerViewRepresentable {
             self.titleLabel.leftAnchor.constraint(equalTo: self.leftAnchor, constant: insets.left),
             self.titleLabel.rightAnchor.constraint(equalTo: self.rightAnchor, constant: insets.right),
         ])
-
-        self.pinBottomOfScrollView(to: self.messageLabel, withPriority: .defaultLow)
-
     }
 
     private func createMessageLabelConstraints() {
@@ -196,7 +209,7 @@ final class AlertView: UIView, AlertControllerViewRepresentable {
             return
         }
 
-        let aligningView = self.textFieldsViewController?.view ?? self.messageLabel!
+        let aligningView = self.textFieldsViewController?.view ?? self.messageLabel
         let widthOffset = self.visualStyle.contentPadding.left + self.visualStyle.contentPadding.right
 
         NSLayoutConstraint.activate([
@@ -233,7 +246,6 @@ final class AlertView: UIView, AlertControllerViewRepresentable {
         let heightConstraint = self.scrollView.heightAnchor.constraint(equalToConstant: height)
         heightConstraint.priority = .defaultHigh
         heightConstraint.isActive = true
-
     }
 
     private func pinBottomOfScrollView(to view: UIView, withPriority priority: UILayoutPriority) {