+++ Source/Views/UIView+Accessibility.swift
@@ -1,9 +1,10 @@
 import UIKit
 
+@available(iOSApplicationExtension, unavailable)
 extension UIView {
     func setupAccessibility(using action: AlertAction) {
         self.accessibilityLabel = action.attributedTitle?.string
-        self.accessibilityTraits = UIAccessibilityTraitButton
+        self.accessibilityTraits = .button
         self.accessibilityIdentifier = action.accessibilityIdentifier
         self.isAccessibilityElement = true
     }